faz casos de teste ex3

+++ src/main/java/ListaExerciciosHerancaPolimorfismo/Ex3/ProdutoEletronico.java
@@ -2,7 +2,7 @@ package ListaExerciciosHerancaPolimorfismo.Ex3;
 
 public class ProdutoEletronico extends Produto {
 
-    public ProdutoEletronico(String nome, double precoBase){
+    public ProdutoEletronico(String nome, float precoBase){
         super(nome, precoBase);
     }
