+++ src/main/java/ListaExerciciosHerancaPolimorfismo/Ex3/ProdutoLivro.java
@@ -1,7 +1,7 @@
 package ListaExerciciosHerancaPolimorfismo.Ex3;
 
 public class ProdutoLivro extends Produto {
-    public ProdutoLivro(String nome, double precoBase){
+    public ProdutoLivro(String nome, float precoBase){
         super(nome, precoBase);
     }
